requirements_lumi.txt
# Requirements specific for LUMI (AMD GPU)
# Install AFTER loading appropriate LUMI modules

# NOTE: On LUMI, first install PyTorch via system modules:
# module load LUMI/24.03 partition/G pytorch/2.4.0

# Then install these packages with pip:

# Lightning for training
pytorch-lightning>=1.6.0
lightning-fabric

# Scientific computing
numpy>=1.20.0
xarray>=0.18.0
netcdf4>=1.5.0

# Data processing
xesmf>=0.6.0
dask[complete]>=2022.0.0
pandas>=1.3.0

# Model utilities
timm>=0.6.0

# Configuration and utilities
click>=8.0.0
tqdm>=4.60.0
pyyaml>=6.0

# For climate data
cftime>=1.5.0
cf-xarray>=0.6.0

# Performance (optional)
numba>=0.55.0

# Development (optional)
jupyter
ipykernel

# Testing (optional)
pytest>=6.0.0

# Note: Avoid installing torch/torchvision via pip on LUMI
# Use system modules that are optimized for AMD GPU
